Supervisord check and tests
This adds a new agent check for supervisord. The check reports whether
supervisord is running or not and  whether one or more processes are
runnnig or not. It also reports prcesses' uptime. The check works over
http inet server and works with sockets. See supervisord.yaml.example
for more details.